# docs/openapi.yaml
openapi: '3.0.2'
info:
    title: Collective IAM
    version: '1.0.0'
servers:
    - url: https://ciam.centralmind.net
paths:
    /role:
        post:
            summary: 'Create a role'
            security:
                - bearerAuth:
                      ['ciam.role.create', 'ciam.permission.grant.{FLAGS}']
            description: |
                Create a new role from a name, description, and optionally also its permission flags.<br>
                `ciam.permission.grant.FLAGS` is only checked if permissions are provided, for example if you try to
                create a role with the permissions [`chipware.ban`, `chipware.unban`] then you will need to have the
                permissions [`ciam.permission.grant.chipware.ban`, `ciam.permission.grant.chipware.unban`].
            tags:
                - role
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                name:
                                    type: string
                                    description: 'The name of this role'
                                description:
                                    type: string
                                    description: 'The description of this role'
                                permissions:
                                    type: array
                                    items:
                                        $ref: '#/components/schemas/Flag'
                                    description: 'The permissions of this role'
                            required:
                                - name
                                - description
            responses:
                '200':
                    description: 'Returns the newly created role'
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Role'
                '400':
                    description: 'Invalid body'
                '401':
                    description: 'Invalid token or missing permissions'
                '500':
                    description: 'Failed saving role to database'

        patch:
            summary: 'Update a role'
            security:
                - bearerAuth:
                      [
                          'ciam.role.update.{ROLE_ID}',
                          'ciam.permission.grant.{FLAGS}',
                          'ciam.permission.revoke.{FLAGS}',
                      ]
            tags:
                - role
            requestBody:
                description: |
                    The role to update, omitted fields wont be updated.<br>
                    `ciam.permission.grant|revoke.FLAGS` is only checked if permissions are provided, for example if you try to
                    set the permissions to [`chipware.ban`, `chipware.unban`] on a role that already has the permissions
                    [`chipware.ban`, `chipware.kick`] then you will need to have the permissions
                    [`ciam.permission.grant.chipware.unban`, `ciam.permission.revoke.chipware.kick`].
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                _id:
                                    $ref: '#/components/schemas/ObjectId'
                                    description: 'The id of the role to update'
                                name:
                                    type: string
                                    description: 'New name of this role'
                                description:
                                    type: string
                                    description: 'New description of this role'
                                permissions:
                                    type: array
                                    items:
                                        $ref: '#/components/schemas/Flag'
                                    description: 'New permissions of this role'
                            required:
                                - _id
            responses:
                '200':
                    description: 'Returns the updated role'
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Role'
                '400':
                    description: 'Invalid body'
                '401':
                    description: 'Invalid token or missing permissions'
                '404':
                    description: 'Role not found'
                '500':
                    description: 'Failed saving role to database'
...
    /user:
        post:
            security:
                - bearerAuth:
                      ['ciam.user.create', 'ciam.permission.grant.{FLAGS}']
            summary: 'Create a user'
            description: |
                Create a new role from at least a name.<br>
                `ciam.permission.grant.FLAGS` is only checked if permissions are provided, for example if you try to
                create a role with the permissions [`chipware.ban`, `chipware.unban`] then you will need to have the
                permissions [`ciam.permission.grant.chipware.ban`, `ciam.permission.grant.chipware.unban`].
            tags:
                - user
            requestBody:
                description: 'The user to create'
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                name:
                                    type: string
                                    description: 'The name of this user'
                                roles:
                                    type: array
                                    items:
                                        $ref: '#/components/schemas/ObjectId'
                                    description: 'The role ids this user has'
                                permissions:
                                    type: array
                                    items:
                                        $ref: '#/components/schemas/Flag'
                                    description: 'The permission flags this user has'
                                discord:
                                    type: object
                                    properties:
                                        id:
                                            $ref: '#/components/schemas/DiscordId'
                                            description: 'User discord id'
                            required:
                                - name
            responses:
                '200':
                    description: 'Returns the newly created user'
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/User'
                '400':
                    description: 'Invalid body, missing role given'
                '401':
                    description: 'Invalid token or missing permissions'
                '500':
                    description: 'Failed saving user to database'

        patch:
            summary: 'Update a user'
            security:
                - bearerAuth:
                      [
                          'ciam.user.update.{USER_ID}',
                          'ciam.permission.grant.{FLAGS}',
                          'ciam.permission.revoke.{FLAGS}',
                      ]
            tags:
                - user
            requestBody:
                description: |
                    The user to update, omitted fields wont be updated.<br>
                    `ciam.permission.grant|revoke.FLAGS` is only checked if permissions are provided, for example if you try to
                    set the permissions to [`chipware.ban`, `chipware.unban`] on a user that already has the permissions
                    [`chipware.ban`, `chipware.kick`] then you will need to have the permissions
                    [`ciam.permission.grant.chipware.unban`, `ciam.permission.revoke.chipware.kick`].
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                _id:
                                    $ref: '#/components/schemas/ObjectId'
                                    description: 'The id of the user to update'
                                name:
                                    type: string
                                    description: 'New name of this user'
                                permissions:
                                    type: array
                                    items:
                                        $ref: '#/components/schemas/Flag'
                                    description: 'New permissions of this user'
                                roles:
                                    type: array
                                    items:
                                        $ref: '#/components/schemas/ObjectId'
                                    description: 'New roles of this user'
                            required:
                                - _id
            responses:
                '200':
                    description: 'Returns the updated user'
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/User'
                '400':
                    description: 'Invalid body'
                '401':
                    description: 'Invalid token or missing permissions'
                '404':
                    description: 'User not found'
                '500':
                    description: 'Failed saving user to database'
...
    /permission:
        post:
            security:
                - bearerAuth: ['ciam.permission.create.{FLAG}']
            summary: 'Create a permission'
            tags:
                - permission
            requestBody:
                description: 'The permission to create, a partial of the Permission object'
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                name:
                                    type: string
                                    description: 'A pretty name for this permission'
                                description:
                                    type: string
                                    description: 'A description for this permission flag'
                                flag:
                                    $ref: '#/components/schemas/FlagStrict'
                                    description: 'The permission flag to create'
                            required:
                                - name
                                - description
                                - flag
            responses:
                '200':
                    description: 'Returns the newly created permission'
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Permission'
                '400':
                    description: 'Bad request'
                '401':
                    description: 'Invalid token or missing permissions'
                '500':
                    description: 'Error when saving the permission to the database'

        patch:
            security:
                - bearerAuth: ['ciam.permission.update.{FLAG}']
            summary: 'Update a permission'
            tags:
                - permission
            requestBody:
                description: 'The permission to update.'
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                flag:
                                    $ref: '#/components/schemas/FlagStrict'
                                    description: 'The flag to update, fails if this flag does not exist'
                                name:
                                    type: string
                                    description: 'The new pretty name of this flag'
                                description:
                                    type: string
                                    description: 'The new description of this flag'
                            required:
                                - flag
            responses:
                '200':
                    description: 'The now modified permission'
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Permission'
                '400':
                    description: 'Invalid body'
                '401':
                    description: 'Invalid token or missing permissions'
                '500':
                    description: 'Failed saving permission to database'
...
    /login:
        get:
            summary: 'Redirect endpoint for OAuth'
            description: |
                Once the user has completed the Oauth2 flow they will be redirected to the `returnUrl`, they will have a cookie
                set for the domain `domain`, this cookie can be used to access the users information.
            tags:
                - auth
            parameters:
                - in: query
                  name: returnUrl
                  description: 'The url to return to after the user has completed oauth'
                  required: true
                  schema:
                      # TODO: Create a schema for urls?
                      type: string
                - in: query
                  name: domain
                  description: 'The domain to set the auth cookie for, this should be your websites domain'
                  required: true
                  schema:
                      # TODO: Create a schema for urls?
                      type: string
                - in: query
                  name: scope
                  description: 'The scope you are requesting for a user, currently only "identify"'
                  required: true
                  schema:
                      # TODO: Create a schema for scopes?
                      type: string
                - in: query
                  name: client
                  description: 'Client id of your application'
                  required: true
                  schema:
                      $ref: '#/components/schemas/ObjectId'
            responses:
                '302':
                    description: 'Redirect to discord OAuth2'
                '400':
                    description: 'Bad request'
    /auth/user:
        get:
            summary: 'Get user'
            description: |
                Get a user from the access token.
            tags:
                - auth
            parameters:
                - in: header
                  name: ciam_token
                  description: 'The access token'
                  required: true
                  schema:
                      type: string
            responses:
                '200':
                    description: 'Redirect to discord OAuth2'
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    id:
                                        $ref: '#/components/schemas/ObjectId'
                                        description: 'User id'
                                    username:
                                        type: string
                                        description: 'Username'
                                    discord:
                                        $ref: '#/components/schemas/UserDiscord'
                '400':
                    description: 'Bad request'
                '401':
                    description: 'Invalid token or missing permissions'
...
components:
    securitySchemes:
        bearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
    schemas:
...
        ObjectId:
            type: string
            maxLength: 24
            pattern: '^[0-9a-f]{24}$'
        DiscordId:
            type: string
            minLength: 16
            maxLength: 20
            pattern: '^[0-9]{16,20}$'
        Flag:
            type: string
            minLength: 1
            pattern: '^(([a-z]+|\?)(\.([a-z]+|\?))*(\.\*)?)|\*$'
        FlagStrict:
            type: string
            minLength: 1
            pattern: '^[a-z0-9]+(\.[a-z0-9]+)*$'
        Role:
            type: object
            description: 'A role is a collection of permission flags, and can be assigned to users'
            properties:
                _id:
                    $ref: '#/components/schemas/ObjectId'
                    description: 'The roles ID, no tthe same as a discord ID'
                name:
                    type: string
                    description: 'The name of the role'
                description:
                    type: string
                    description: 'A description of this role'
                permissions:
                    type: array
                    description: 'A list of all the permission flags this role has'
                    items:
                        $ref: '#/components/schemas/Flag'
                creator:
                    type: string
                    description: 'The user id of the user who created this role'
            required:
                - _id
                - name
                - description
                - permissions
                - creator
        UserDiscord:
            type: object
            description: 'User discord data'
            properties:
                id:
                    $ref: '#/components/schemas/DiscordId'
                    description: 'Discord id'
                name:
                    type: string
                    description: 'Discord username'
                discriminator:
                    type: string
                    description: 'Discord discriminator'
            required:
                - id
        User:
            type: object
            description: 'A user can have both permissions and roles, and can be tied to a discord user, or be standalone'
            properties:
                _id:
                    $ref: '#/components/schemas/ObjectId'
                    description: 'The users ID, not the same as a discord ID'
                name:
                    type: string
                    description: 'The name of this user'
                avatar:
                    type: string
                    description: 'Avatar url'
                roles:
                    type: array
                    items:
                        $ref: '#/components/schemas/ObjectId'
                    description: 'List of role ids this user has'
                permissions:
                    type: array
                    items:
                        $ref: '#/components/schemas/Flag'
                    description: 'List of permission flags this user has'
                discord:
                    $ref: '#/components/schemas/UserDiscord'
            required:
                - _id
                - name
                - avatar
                - roles
                - permissions
        Permission:
            type: object
            description: 'A permission has a string flag, and a description'
            properties:
                name:
                    type: string
                    description: 'A pretty name for a description'
                description:
                    type: string
                    description: 'A description for this specific permission flag'
                key:
                    type: string
                    description: 'The last path in this permission flag, this is what the description is actually describing'
                path:
                    type: array
                    items:
                        type: string
                    description: 'The path to this permission key'
                flag:
                    $ref: '#/components/schemas/FlagStrict'
                    description: 'The path and key combined'
                creator:
                    $ref: '#/components/schemas/ObjectId'
                    description: 'The user who created this permission'
            required:
                - name
                - description
                - key
                - path
                - flag
                - creator
...